Optional: create/use defines for I2C Address
Not completely necessary, but it took me awhile to figure out why I was
not seeing something on my DIYMALL display.  Finally figured out that it
is soldered to set address to 0x37 (marked 78).

So put a couple of defines in with comments for 3C and 3D in ino, and
used the define instead of 0x3D in the display.begin statement