Make resolution selectable at runtime and display buffer private
This makes it possible to drive different types of displays and multiple different displays at once, possibly of different resolutions.
It does increase memory requirements: maximum size buffer (1024 bytes) is allocated, half of which will be wasted if display is 128x32.
Also logo is now in a separate buffer.

Tested with I2C 128x64 and SPI 128x32 displays connected at the same time.