save correct contrast setting for use in dim() function
In begin(), there is code to compute the correct contrast value, based on the configuration of the display.
The same code also computes the comPin setting. This code has been factored so that these two values
are computed, and contrast saved in an instance variable.

Later, in dim(), the saved computed contrast value is used when un-dimming the display.