Added comments for pins used on different boards
To help users getting started less painful. You find 1000 pics of arduino UNOs connected at pins A4, A5 to the display but it doesn't work on your MEGA 2560 because there it's the pins 20, 21 that are used for I2C. You don't necessarily know that already at the moment you start with these displays. And you maybe do not search in the Wire-library for the solution.